Created module for studentstruct
